+++ src/cseval/cseval.cpp
@@ -1,4 +1,4 @@
-#include "cseval.hpp"
+#include "./cseval.hpp"
 
 template <typename Real>
 const Real cseval<Real>::ZERO = Real("0");
@@ -95,7 +95,8 @@ cseval<Real>::cseval(std::string expression, char imaginary_unit)
 #endif
   if (expression.empty()) {
     throw std::invalid_argument(
-        "Expression string is empty or some substring within brackets is empty");
+        "Expression string is empty or some substring within brackets is "
+        "empty");
   }
 
   // Remove braces.
